targets/auth-service/requirements.txt
flask
pyjwt
psycopg2-binary
